requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / LLM
google-generativeai==0.8.3
anthropic==0.40.0

# Data
pandas==2.2.3
numpy==2.1.3
pyarrow==18.0.0

# Utils
pydantic==2.10.1
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.0

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0